MAT-7522: some cleanup, refactors and re-organization; read and modify top-level data on test case bundle entries
