+++ src/components/editTestCase/qiCore/LeftPanel/ElementsTab/builder/element/types/BooleanComponent.tsx
